Added a assignment from uni
Since it was a university project, it is in potuguese, and it uses the webpages i made that are in my other repository, but now it has a php backend. the project was made in group and we didnt do control version in github because time was short. so i am publishing everything with one commit